Contracts refactor phase 6: extract ContractActionPanelComponent
Moves the interact/action dispatcher (interactContract(), all per-function
execution methods, two-phase/partial-spend signing) and both action-form
template regions (curated full-page form + advanced/manual ABI-driven
fallback) into a new standalone component.

Interact-state (interactContractJson, outpoint/amount, selectedFunction,
etc.) stays shell-owned via two-way model() bindings rather than moving
outright — tracing every read/write turned up several call sites still
living in not-yet-extracted Phase 7 code (selectDetailAction,
prepareDetailInteractState, selectDefaultFunctionForContract,
clearInteractContractSelection, restoreTransientState) that write into it.
A new pendingFunctionSelect command channel replaces the shell's previous
direct calls into selectFunction(), which now lives on the panel.

Registry updates and indexer-status polling are relayed back to the shell
via registryEntryUpdated/actionIndexingRequested outputs, reusing the
existing onRegistryEntryUpdated handler from the deploy-form extraction.
