+++ src/app/v2/pages/app/flows/contracts/components/contract-action-panel/contract-action-panel.component.scss
@@ -1,0 +1,130 @@
+// Duplicated (not moved) from ../../contracts-page.component.scss — Angular's
+// emulated view encapsulation scopes CSS per-component. .panel, .back-button,
+// .contract-summary, .result-panel, .max-text, .advanced-section, .field*,
+// .banner and .contract-input/.form-input base styles already live in
+// src/styles/shared-ui.scss (global), so only the rules below — which are
+// still local to the shell's own stylesheet — need a copy here.
+
+.alias-actions {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  flex-wrap: wrap;
+  margin-top: 4px;
+}
+
+.alias-link,
+.alias-btn {
+  border: 0;
+  background: transparent;
+  color: var(--kaspa-20);
+  font: inherit;
+  font-size: 12px;
+  padding: 0;
+  cursor: pointer;
+}
+
+.alias-link.muted,
+.alias-btn.ghost {
+  color: var(--gray-60);
+}
+
+.alias-editor {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  flex-wrap: wrap;
+  margin-top: 8px;
+}
+
+.detail-alias-editor {
+  margin-top: 0;
+}
+
+.alias-notice {
+  margin-top: 6px;
+  color: var(--warning-50, #f8c96a);
+  font-size: 12px;
+  line-height: 1.4;
+}
+
+.action-contract-name {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 12px;
+  flex-wrap: wrap;
+  padding: 12px 14px;
+  border: 1px solid var(--border-secondary);
+  border-radius: 10px;
+  background: var(--gray-10);
+}
+
+.alias-input {
+  min-width: 180px;
+  max-width: 100%;
+  height: 32px;
+  border: 1px solid var(--border-secondary);
+  border-radius: 8px;
+  background: var(--gray-15);
+  color: var(--white);
+  padding: 0 10px;
+  font-size: 13px;
+  outline: none;
+
+  &:focus {
+    border-color: var(--kaspa-20);
+  }
+}
+
+.error-message {
+  background: rgba(224, 66, 66, 0.1);
+  border: 1px solid var(--red-20);
+  color: var(--red-20);
+  border-radius: 10px;
+  padding: 12px;
+  font-size: 13px;
+  word-break: break-word;
+}
+
+// .contract-input/.form-input base styles live in src/styles/shared-ui.scss;
+// only the textarea-specific sizing stays local.
+.contract-input {
+  resize: vertical;
+  min-height: 120px;
+  font-size: 13px;
+}
+
+.interact-contract-title {
+  @media (max-width: 767px) {
+    font-size: 10px;
+  }
+}
+
+// kc-button's "m" size variant enforces a 300px min-width on its inner
+// .kc-button element (an ui-kit design-token default sized for standalone
+// CTAs). Wherever a kc-button sits inline next to an input or alongside
+// other buttons in this page (instead of opting into [isFullWidth]), that
+// fixed minimum overflows the row on narrow (320px) viewports. Let those
+// buttons hug their own content instead.
+.flex.gap-8.align-items-end.wrap kc-button,
+.flex.gap-8.wrap kc-button,
+.flex.align-items-center.gap-8 kc-button {
+  ::ng-deep .kc-button {
+    min-width: 0;
+  }
+}
+
+// Active function button highlight — kc-button's host element has no
+// border-radius of its own (the rounding lives on its inner .kc-button),
+// so styling the host directly draws a sharp-cornered box around a rounded
+// button. Target the inner element instead so the highlight follows the
+// button's actual corners.
+.active-function {
+  ::ng-deep .kc-button {
+    border-color: var(--kaspa-20) !important;
+    background: rgba(111, 199, 186, 0.22) !important;
+    box-shadow: 0 0 0 2px rgba(111, 199, 186, 0.18) !important;
+    color: #fff !important;
+  }
+}
